+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/node/Node.java
@@ -45,9 +45,6 @@ public abstract class Node implements UniqueId {
   /** Is this node an l-value? */
   protected boolean lvalue = false;
 
-  /** The assignment context of this node. See {@link AssignmentContext}. */
-  protected @Nullable AssignmentContext assignmentContext;
-
   /**
    * Does this node represent a tree that appears in the source code (true) or one that the CFG
    * builder added while desugaring (false).
@@ -146,15 +143,6 @@ public abstract class Node implements UniqueId {
 
   public void setInSource(boolean inSrc) {
     inSource = inSrc;
-  }
-
-  /** The assignment context for the node. */
-  public @Nullable AssignmentContext getAssignmentContext() {
-    return assignmentContext;
-  }
-
-  public void setAssignmentContext(AssignmentContext assignmentContext) {
-    this.assignmentContext = assignmentContext;
   }
 
   /**