+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -5,7 +5,6 @@ import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.printer.DefaultPrettyPrinter;
 import com.sun.source.tree.AnnotatedTypeTree;
 import com.sun.source.tree.AnnotationTree;
-import com.sun.source.tree.ArrayAccessTree;
 import com.sun.source.tree.ArrayTypeTree;
 import com.sun.source.tree.AssignmentTree;
 import com.sun.source.tree.CatchTree;
@@ -121,7 +120,6 @@ import org.checkerframework.framework.type.AnnotatedTypeParameterBounds;
 import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
 import org.checkerframework.framework.type.QualifierHierarchy;
 import org.checkerframework.framework.type.TypeHierarchy;
-import org.checkerframework.framework.type.VisitorState;
 import org.checkerframework.framework.type.poly.QualifierPolymorphism;
 import org.checkerframework.framework.type.visitor.SimpleAnnotatedTypeScanner;
 import org.checkerframework.framework.util.AnnotatedTypes;
@@ -178,15 +176,6 @@ import org.plumelib.util.CollectionsPlume;
  * @see TypeHierarchy#isSubtype(AnnotatedTypeMirror, AnnotatedTypeMirror)
  * @see AnnotatedTypeFactory
  */
-/*
- * Note how the handling of VisitorState is duplicated in AbstractFlow. In
- * particular, the handling of the assignment context has to be done correctly
- * in both classes. This is a pain and we should see how to handle this in the
- * DFF version.
- *
- * TODO: missing assignment context: array initializer
- * expressions should have the component type as context
- */
 public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?, ?>>
     extends SourceVisitor<Void, Void> {
 
@@ -198,9 +187,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
   /** For obtaining line numbers in -Ashowchecks debugging output. */
   protected final SourcePositions positions;
-
-  /** For storing visitor state. */
-  protected final VisitorState visitorState;
 
   /** The element for java.util.Vector#copyInto. */
   private final ExecutableElement vectorCopyInto;
@@ -250,6 +236,9 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
    */
   protected boolean inferPurity = true;
 
+  /** The tree of the enclosing method that is currently being visited. */
+  protected @Nullable MethodTree methodTree = null;
+
   /**
    * @param checker the type-checker associated with this visitor (for callbacks to {@link
    *     TypeHierarchy#isSubtype})
@@ -268,7 +257,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     this.checker = checker;
     this.atypeFactory = typeFactory == null ? createTypeFactory() : typeFactory;
     this.positions = trees.getSourcePositions();
-    this.visitorState = atypeFactory.getVisitorState();
     this.typeValidator = createTypeValidator();
     ProcessingEnvironment env = checker.getProcessingEnvironment();
     this.vectorCopyInto = TreeUtils.getMethod("java.util.Vector", "copyInto", 1, env);
@@ -351,7 +339,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
   @Override
   public Void scan(@Nullable Tree tree, Void p) {
     if (tree != null && getCurrentPath() != null) {
-      this.visitorState.setPath(new TreePath(getCurrentPath(), tree));
+      this.atypeFactory.setVisitorTreePath(new TreePath(getCurrentPath(), tree));
     }
     return super.scan(tree, p);
   }
@@ -481,32 +469,19 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     }
     atypeFactory.preProcessClassTree(classTree);
 
-    TreePath preTreePath = visitorState.getPath();
-    AnnotatedDeclaredType preACT = visitorState.getClassType();
-    ClassTree preCT = visitorState.getClassTree();
-    AnnotatedDeclaredType preAMT = visitorState.getMethodReceiver();
-    MethodTree preMT = visitorState.getMethodTree();
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
+    TreePath preTreePath = atypeFactory.getVisitorTreePath();
+    MethodTree preMT = methodTree;
 
-    // Don't use atypeFactory.getPath, because that depends on the visitorState path.
-    visitorState.setPath(TreePath.getPath(root, classTree));
-    visitorState.setClassType(
-        atypeFactory.getAnnotatedType(TreeUtils.elementFromDeclaration(classTree)));
-    visitorState.setClassTree(classTree);
-    visitorState.setMethodReceiver(null);
-    visitorState.setMethodTree(null);
-    visitorState.setAssignmentContext(null);
+    // Don't use atypeFactory.getPath, because that depends on the visitor path.
+    atypeFactory.setVisitorTreePath(TreePath.getPath(root, classTree));
+    methodTree = null;
 
     try {
       processClassTree(classTree);
       atypeFactory.postProcessClassTree(classTree);
     } finally {
-      visitorState.setPath(preTreePath);
-      visitorState.setClassType(preACT);
-      visitorState.setClassTree(preCT);
-      visitorState.setMethodReceiver(preAMT);
-      visitorState.setMethodTree(preMT);
-      visitorState.setAssignmentContext(preAssignmentContext);
+      atypeFactory.setVisitorTreePath(preTreePath);
+      methodTree = preMT;
     }
     return null;
   }
@@ -891,10 +866,8 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     // later checks.
     // TODO: Find a cleaner way to ensure circular AnnotatedTypeMirrors.
     AnnotatedExecutableType methodType = atypeFactory.getAnnotatedType(node).deepCopy();
-    AnnotatedDeclaredType preMRT = visitorState.getMethodReceiver();
-    MethodTree preMT = visitorState.getMethodTree();
-    visitorState.setMethodReceiver(methodType.getReceiverType());
-    visitorState.setMethodTree(node);
+    MethodTree preMT = methodTree;
+    methodTree = node;
     ExecutableElement methodElement = TreeUtils.elementFromDeclaration(node);
 
     warnAboutTypeAnnotationsTooEarly(node, node.getModifiers());
@@ -976,8 +949,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
       return super.visitMethod(node, p);
     } finally {
-      visitorState.setMethodReceiver(preMRT);
-      visitorState.setMethodTree(preMT);
+      methodTree = preMT;
     }
   }
 
@@ -1442,7 +1414,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
     visitAnnotatedType(node.getModifiers().getAnnotations(), node.getType());
 
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
     AnnotatedTypeMirror variableType;
     if (getCurrentPath().getParentPath() != null
         && getCurrentPath().getParentPath().getLeaf().getKind() == Tree.Kind.LAMBDA_EXPRESSION) {
@@ -1453,22 +1424,17 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     } else {
       variableType = atypeFactory.getAnnotatedTypeLhs(node);
     }
-    visitorState.setAssignmentContext(Pair.of(node, variableType));
 
-    try {
-      atypeFactory.getDependentTypesHelper().checkTypeForErrorExpressions(variableType, node);
-      // If there's no assignment in this variable declaration, skip it.
-      if (node.getInitializer() != null) {
-        commonAssignmentCheck(node, node.getInitializer(), "assignment");
-      } else {
-        // commonAssignmentCheck validates the type of node,
-        // so only validate if commonAssignmentCheck wasn't called
-        validateTypeOf(node);
-      }
-      return super.visitVariable(node, p);
-    } finally {
-      visitorState.setAssignmentContext(preAssignmentContext);
+    atypeFactory.getDependentTypesHelper().checkTypeForErrorExpressions(variableType, node);
+    // If there's no assignment in this variable declaration, skip it.
+    if (node.getInitializer() != null) {
+      commonAssignmentCheck(node, node.getInitializer(), "assignment");
+    } else {
+      // commonAssignmentCheck validates the type of node,
+      // so only validate if commonAssignmentCheck wasn't called
+      validateTypeOf(node);
     }
+    return super.visitVariable(node, p);
   }
 
   /**
@@ -1594,15 +1560,8 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
    */
   @Override
   public Void visitAssignment(AssignmentTree node, Void p) {
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
-    visitorState.setAssignmentContext(
-        Pair.of((Tree) node.getVariable(), atypeFactory.getAnnotatedType(node.getVariable())));
-    try {
-      commonAssignmentCheck(node.getVariable(), node.getExpression(), "assignment");
-      return super.visitAssignment(node, p);
-    } finally {
-      visitorState.setAssignmentContext(preAssignmentContext);
-    }
+    commonAssignmentCheck(node.getVariable(), node.getExpression(), "assignment");
+    return super.visitAssignment(node, p);
   }
 
   /**
@@ -1997,7 +1956,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
       // Check return type for single statement returns here.
       AnnotatedTypeMirror ret = functionType.getReturnType();
       if (ret.getKind() != TypeKind.VOID) {
-        visitorState.setAssignmentContext(Pair.of((Tree) node, ret));
         commonAssignmentCheck(ret, (ExpressionTree) node.getBody(), "return");
       }
     }
@@ -2037,37 +1995,29 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
       return super.visitReturn(node, p);
     }
 
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
-    try {
+    Tree enclosing =
+        TreePathUtil.enclosingOfKind(
+            getCurrentPath(),
+            new HashSet<>(Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION)));
 
-      Tree enclosing =
-          TreePathUtil.enclosingOfKind(
-              getCurrentPath(),
-              new HashSet<>(Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION)));
+    AnnotatedTypeMirror ret = null;
+    if (enclosing.getKind() == Tree.Kind.METHOD) {
 
-      AnnotatedTypeMirror ret = null;
-      if (enclosing.getKind() == Tree.Kind.METHOD) {
-
-        MethodTree enclosingMethod = TreePathUtil.enclosingMethod(getCurrentPath());
-        boolean valid = validateTypeOf(enclosing);
-        if (valid) {
-          ret = atypeFactory.getMethodReturnType(enclosingMethod, node);
-        }
-      } else {
-        AnnotatedExecutableType result =
-            atypeFactory.getFunctionTypeFromTree((LambdaExpressionTree) enclosing);
-        ret = result.getReturnType();
+      MethodTree enclosingMethod = TreePathUtil.enclosingMethod(getCurrentPath());
+      boolean valid = validateTypeOf(enclosing);
+      if (valid) {
+        ret = atypeFactory.getMethodReturnType(enclosingMethod, node);
       }
-
-      if (ret != null) {
-        visitorState.setAssignmentContext(Pair.of((Tree) node, ret));
-
-        commonAssignmentCheck(ret, node.getExpression(), "return");
-      }
-      return super.visitReturn(node, p);
-    } finally {
-      visitorState.setAssignmentContext(preAssignmentContext);
+    } else {
+      AnnotatedExecutableType result =
+          atypeFactory.getFunctionTypeFromTree((LambdaExpressionTree) enclosing);
+      ret = result.getReturnType();
     }
+
+    if (ret != null) {
+      commonAssignmentCheck(ret, node.getExpression(), "return");
+    }
+    return super.visitReturn(node, p);
   }
 
   /**
@@ -2128,37 +2078,20 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
       }
 
       AnnotatedTypeMirror expected = annoTypes.get(at.getVariable().toString());
-      Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
-
-      {
-        // Determine and set the new assignment context.
-        ExpressionTree var = at.getVariable();
-        assert var instanceof IdentifierTree : "Expected IdentifierTree as context. Found: " + var;
-        AnnotatedTypeMirror meth = atypeFactory.getAnnotatedType(var);
-        assert meth instanceof AnnotatedExecutableType
-            : "Expected AnnotatedExecutableType as context. Found: " + meth;
-        AnnotatedTypeMirror newctx = ((AnnotatedExecutableType) meth).getReturnType();
-        visitorState.setAssignmentContext(Pair.of((Tree) null, newctx));
-      }
-
-      try {
-        AnnotatedTypeMirror actual = atypeFactory.getAnnotatedType(at.getExpression());
-        if (expected.getKind() != TypeKind.ARRAY) {
-          // Expected is not an array -> direct comparison.
-          commonAssignmentCheck(expected, actual, at.getExpression(), "annotation");
-        } else if (actual.getKind() == TypeKind.ARRAY) {
-          // Both actual and expected are arrays.
-          commonAssignmentCheck(expected, actual, at.getExpression(), "annotation");
-        } else {
-          // The declaration is an array type, but just a single element is given.
-          commonAssignmentCheck(
-              ((AnnotatedArrayType) expected).getComponentType(),
-              actual,
-              at.getExpression(),
-              "annotation");
-        }
-      } finally {
-        visitorState.setAssignmentContext(preAssignmentContext);
+      AnnotatedTypeMirror actual = atypeFactory.getAnnotatedType(at.getExpression());
+      if (expected.getKind() != TypeKind.ARRAY) {
+        // Expected is not an array -> direct comparison.
+        commonAssignmentCheck(expected, actual, at.getExpression(), "annotation");
+      } else if (actual.getKind() == TypeKind.ARRAY) {
+        // Both actual and expected are arrays.
+        commonAssignmentCheck(expected, actual, at.getExpression(), "annotation");
+      } else {
+        // The declaration is an array type, but just a single element is given.
+        commonAssignmentCheck(
+            ((AnnotatedArrayType) expected).getComponentType(),
+            actual,
+            at.getExpression(),
+            "annotation");
       }
     }
     return null;
@@ -2432,19 +2365,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
       }
     }
     return super.visitInstanceOf(node, p);
-  }
-
-  @Override
-  public Void visitArrayAccess(ArrayAccessTree node, Void p) {
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
-    try {
-      visitorState.setAssignmentContext(null);
-      scan(node.getExpression(), p);
-      scan(node.getIndex(), p);
-    } finally {
-      visitorState.setAssignmentContext(preAssignmentContext);
-    }
-    return null;
   }
 
   /**
@@ -3289,23 +3209,16 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             executableName,
             listToString(paramNames));
 
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext = visitorState.getAssignmentContext();
-    try {
-      for (int i = 0; i < size; ++i) {
-        visitorState.setAssignmentContext(
-            Pair.of((Tree) null, (AnnotatedTypeMirror) requiredArgs.get(i)));
-        commonAssignmentCheck(
-            requiredArgs.get(i),
-            passedArgs.get(i),
-            "argument",
-            // TODO: for expanded varargs parameters, maybe adjust the name
-            paramNames.get(Math.min(i, maxParamNamesIndex)),
-            executableName);
-        // Also descend into the argument within the correct assignment context.
-        scan(passedArgs.get(i), null);
-      }
-    } finally {
-      visitorState.setAssignmentContext(preAssignmentContext);
+    for (int i = 0; i < size; ++i) {
+      commonAssignmentCheck(
+          requiredArgs.get(i),
+          passedArgs.get(i),
+          "argument",
+          // TODO: for expanded varargs parameters, maybe adjust the name
+          paramNames.get(Math.min(i, maxParamNamesIndex)),
+          executableName);
+      // Also descend into the argument within the correct assignment context.
+      scan(passedArgs.get(i), null);
     }
   }
 
@@ -4137,7 +4050,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
    * expression in {@code mustSubset} there must be the same expression in {@code set}, with the
    * same (or a stronger) annotation.
    *
-   * <p>This uses field {@link #visitorState} to determine where to issue an error message.
+   * <p>This uses field {@link #methodTree} to determine where to issue an error message.
    *
    * @param overriderType the subtype
    * @param overriddenType the supertype
@@ -4167,7 +4080,6 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
       }
 
       if (!found) {
-        MethodTree methodDeclTree = visitorState.getMethodTree();
 
         String overriddenTypeString = overriddenType.getUnderlyingType().asElement().toString();
         String overriderTypeString;
@@ -4201,10 +4113,10 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
         }
 
         checker.reportError(
-            methodDeclTree,
+            methodTree,
             messageKey,
             weak.first,
-            methodDeclTree.getName(),
+            methodTree.getName(),
             overriddenTypeString,
             overriddenAnno,
             overriderTypeString,
@@ -4215,15 +4127,15 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
   /**
    * Localizes some contracts -- that is, viewpoint-adapts them to some method body, according to
-   * the value of {@link #visitorState}.
+   * the value of {@link #methodTree}.
    *
    * <p>The input is a set of {@link Contract}s, each of which contains an expression string and an
    * annotation. In a {@link Contract}, Java expressions are exactly as written in source code, not
    * standardized or viewpoint-adapted.
    *
    * <p>The output is a set of pairs of {@link JavaExpression} (parsed expression string) and
-   * standardized annotation (with respect to the path of {@link #visitorState}. This method
-   * discards any contract whose expression cannot be parsed into a JavaExpression.
+   * standardized annotation (with respect to the path of {@link #methodTree}. This method discards
+   * any contract whose expression cannot be parsed into a JavaExpression.
    *
    * @param contractSet a set of contracts
    * @param methodType the type of the method that the contracts are for
@@ -4238,7 +4150,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
     // This is the path to a place where the contract is being used, which might or might not be
     // where the contract was defined.  For example, methodTree might be an overriding
     // definition, and the contract might be for a superclass.
-    MethodTree methodTree = visitorState.getMethodTree();
+    MethodTree methodTree = this.methodTree;
 
     StringToJavaExpression stringToJavaExpr =
         expression -> {