+++ framework/src/main/java/org/checkerframework/framework/util/AnnotatedTypes.java
@@ -1018,10 +1018,13 @@ public class AnnotatedTypes {
    * Return a list of the AnnotatedTypeMirror of the passed expression trees, in the same order as
    * the trees.
    *
+   * @param atypeFactory a type factory
    * @param paramTypes the parameter types to use as assignment context
    * @param trees the AST nodes
    * @return a list with the AnnotatedTypeMirror of each tree in trees
+   * @deprecated use CollectionsPlume.mapList(atypeFactory::getAnnotatedType, trees) instead.
    */
+  @Deprecated
   public static List<AnnotatedTypeMirror> getAnnotatedTypes(
       AnnotatedTypeFactory atypeFactory,
       List<AnnotatedTypeMirror> paramTypes,
@@ -1034,21 +1037,8 @@ public class AnnotatedTypes {
               + " Arguments: "
               + trees);
     }
-    Pair<Tree, AnnotatedTypeMirror> preAssignmentContext =
-        atypeFactory.getVisitorState().getAssignmentContext();
 
-    List<AnnotatedTypeMirror> types = new ArrayList<>();
-    try {
-      for (int i = 0; i < trees.size(); ++i) {
-        AnnotatedTypeMirror param = paramTypes.get(i);
-        atypeFactory.getVisitorState().setAssignmentContext(Pair.of((Tree) null, param));
-        ExpressionTree arg = trees.get(i);
-        types.add(atypeFactory.getAnnotatedType(arg));
-      }
-    } finally {
-      atypeFactory.getVisitorState().setAssignmentContext(preAssignmentContext);
-    }
-    return types;
+    return CollectionsPlume.mapList(atypeFactory::getAnnotatedType, trees);
   }
 
   /**