+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractTransfer.java
@@ -74,7 +74,6 @@ import org.checkerframework.framework.util.ContractsFromMethod;
 import org.checkerframework.framework.util.JavaExpressionParseUtil.JavaExpressionParseException;
 import org.checkerframework.framework.util.StringToJavaExpression;
 import org.checkerframework.javacutil.ElementUtils;
-import org.checkerframework.javacutil.Pair;
 import org.checkerframework.javacutil.TreePathUtil;
 import org.checkerframework.javacutil.TreeUtils;
 
@@ -189,36 +188,7 @@ public abstract class CFAbstractTransfer<
     GenericAnnotatedTypeFactory<V, S, T, ? extends CFAbstractAnalysis<V, S, T>> factory =
         analysis.atypeFactory;
     Tree preTree = analysis.getCurrentTree();
-    Pair<Tree, AnnotatedTypeMirror> preContext = factory.getVisitorState().getAssignmentContext();
     analysis.setCurrentTree(tree);
-    // is there an assignment context node available?
-    if (node != null && node.getAssignmentContext() != null) {
-      // Get the declared type of the assignment context by looking up the assignment context tree's
-      // type in the factory while flow is disabled.
-      Tree contextTree = node.getAssignmentContext().getContextTree();
-      AnnotatedTypeMirror assignmentContext = null;
-      if (contextTree != null) {
-        assignmentContext = factory.getAnnotatedTypeLhs(contextTree);
-      } else {
-        Element assignmentContextElement = node.getAssignmentContext().getElementForType();
-        if (assignmentContextElement != null) {
-          // if contextTree is null, use the element to get the type
-          assignmentContext = factory.getAnnotatedType(assignmentContextElement);
-        }
-      }
-
-      if (assignmentContext != null) {
-        if (assignmentContext instanceof AnnotatedExecutableType) {
-          // For a MethodReturnContext, we get the full type of the
-          // method, but we only want the return type.
-          assignmentContext = ((AnnotatedExecutableType) assignmentContext).getReturnType();
-        }
-        factory
-            .getVisitorState()
-            .setAssignmentContext(
-                Pair.of(node.getAssignmentContext().getContextTree(), assignmentContext));
-      }
-    }
     AnnotatedTypeMirror at;
     if (node instanceof MethodInvocationNode
         && ((MethodInvocationNode) node).getIterableExpression() != null) {
@@ -232,7 +202,6 @@ public abstract class CFAbstractTransfer<
       at = factory.getAnnotatedType(tree);
     }
     analysis.setCurrentTree(preTree);
-    factory.getVisitorState().setAssignmentContext(preContext);
     return analysis.createAbstractValue(at);
   }