+++ docs/CHANGELOG.md
@@ -5,6 +5,12 @@ Version 3.1?.? (November 1, 2021)
 
 **Implementation details:**
 
+Removed `org.checkerframework.framework.type.VisitorState`
+
+Deprecated methods in AnnotatedTypeFactory:
+* `getCurrentClassTree`
+* `getCurrentMethodReceiver`
+
 **Closed issues:**
 
 Version 3.18.1 (October 4, 2021)