+++ docs/CHANGELOG.md
@@ -1,3 +1,18 @@
+Version 3.1?.? (November 1, 2021)
+-------------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+Removed `org.checkerframework.framework.type.VisitorState`
+
+Deprecated methods in AnnotatedTypeFactory:
+* `getCurrentClassTree`
+* `getCurrentMethodReceiver`
+
+**Closed issues:**
+
 Version 3.18.1 (October 4, 2021)
 -------------------------------